bugfix:fix codex bug

+++ services/location_locator/locator/source/network_ability_proxy.cpp
@@ -132,7 +132,7 @@ bool NetworkAbilityProxy::SetMocked(
         return false;
     }
     config.Marshalling(data);
-    int locationSize = location.size();
+    int locationSize = static_cast<int>(location.size());
     data.WriteInt32(locationSize);
     for (int i = 0; i < locationSize; i++) {
         location.at(i)->Marshalling(data);
